TIDEMAN Strategy
Tideman strategy to the best of my ability (differs in that TIDEMAN has a condition where it'll forigve opponent if number of defections differs from a random generator by 3 standard deviations. Unclear if some calculation is done at runtime or if the value was determined before runtime and hard coded).